Generates C files with Cython
Note that these are pretty much unreadable (to me anyway) so they might
not be too helpful. However if you want faster code or want a small
script to be written in C, it is a good idea to use Cython. You can also
do this on the interpreter level with the Jupyter QtConsole. I think I
have also read that Cython can be used as a superset to both C and
Python, but I don't know the details. All of those files can be found in
the [/../cython] directory.

In addition to DS, which is very memory heavy, HelperMethods also includes
a new findClosestNode method which is vectorized and far quicker than the
naive method found in some other files.

I'm also having some problems with the AStar algorithm in that it is
actually far slower than Dijkstra. If someone could figure out why that
is happening that would be positively swell. Pretty sure it has to do
with the heuristic, since that's literally the only line of code
changed between the two.

Also, to actually use DS, you have to have a list of tuples, pick an
order of descent (usually no more than 1) and choose your bounds. Then
use the prepare function to obtain a hashtable (hashmap, dictionary,
etc...). After this, you can find the closest node extremely quickly
with the desselect function which requires the hashtable, order, the
point that you are looking at, and the bounds.

More notes:
The keyword [pass] specifically does nothing in Python, I just use it to
denote the end of code blocks. You do not need to do that at all, but
personally it helps with readability.

QTPathFinder is still very fragile, so expect bugs.

I'm going to take a short break from working on this to finish other
things, so if people could wrap it up that would make me very happy.