Logic fixes and handling for memory overflow
For higher threshold values memory overflow is a problem, I added some
control variables to prevent this and some handling for when they are
saving memory too conservatively.  We still need some handling for when
they overestimate, allowing memory overflow to happen.
Added logic that allows paths to form between diagonally adjacent nodes.
If we don't want this it is simple to remove.